Move Loggable to Traits directory; use trait boot method for observers
