+++ app/Models/Component.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Models\Traits\Loggable;
 use App\Models\Traits\Searchable;
 use App\Presenters\Presentable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
