+++ app/Observers/AssetObserver.php
@@ -10,63 +10,6 @@ use Carbon\Carbon;
 
 class AssetObserver
 {
-    /**
-     * Listen to the Asset updating event. This fires automatically every time an existing asset is saved.
-     *
-     * @param  Asset  $asset
-     * @return void
-     */
-    public function updating(Asset $asset)
-    {
-        $attributes = $asset->getAttributes();
-        $attributesOriginal = $asset->getRawOriginal();
-        $same_checkout_counter = false;
-        $same_checkin_counter = false;
-        $restoring_or_deleting = false;
-
-
-        // This is a gross hack to prevent the double logging when restoring an asset
-        if (array_key_exists('deleted_at', $attributes)  && array_key_exists('deleted_at', $attributesOriginal)){
-            $restoring_or_deleting = (($attributes['deleted_at'] != $attributesOriginal['deleted_at']));
-        }
-
-        if (array_key_exists('checkout_counter', $attributes) && array_key_exists('checkout_counter', $attributesOriginal)){
-            $same_checkout_counter = (($attributes['checkout_counter'] == $attributesOriginal['checkout_counter']));
-        }
-
-        if (array_key_exists('checkin_counter', $attributes)  && array_key_exists('checkin_counter', $attributesOriginal)){
-            $same_checkin_counter = (($attributes['checkin_counter'] == $attributesOriginal['checkin_counter']));
-        }
-
-        // If the asset isn't being checked out or audited, log the update.
-        // (Those other actions already create log entries.)
-	    if (($attributes['assigned_to'] == $attributesOriginal['assigned_to'])
-	    && ($same_checkout_counter) && ($same_checkin_counter)
-            && ((isset( $attributes['next_audit_date']) ? $attributes['next_audit_date'] : null) == (isset($attributesOriginal['next_audit_date']) ? $attributesOriginal['next_audit_date']: null))
-            && ($attributes['last_checkout'] == $attributesOriginal['last_checkout']) && (!$restoring_or_deleting))
-        {
-            $changed = [];
-
-            foreach ($asset->getRawOriginal() as $key => $value) {
-                if ((array_key_exists($key, $asset->getAttributes())) && ($asset->getRawOriginal()[$key] != $asset->getAttributes()[$key])) {
-                    $changed[$key]['old'] = $asset->getRawOriginal()[$key];
-                    $changed[$key]['new'] = $asset->getAttributes()[$key];
-                }
-	    }
-
-	    if (empty($changed)){
-	        return;
-	    }
-
-            $logAction = new Actionlog();
-            $logAction->item_type = Asset::class;
-            $logAction->item_id = $asset->id;
-            $logAction->created_at = date('Y-m-d H:i:s');
-            $logAction->created_by = auth()->id();
-            $logAction->log_meta = json_encode($changed);
-            $logAction->logaction('update');
-        }
-    }
 
     /**
      * Listen to the Asset created event, and increment
@@ -115,37 +58,7 @@ class AssetObserver
         $logAction->logaction('create');
     }
 
-    /**
-     * Listen to the Asset deleting event.
-     *
-     * @param  Asset  $asset
-     * @return void
-     */
-    public function deleting(Asset $asset)
-    {
-        $logAction = new Actionlog();
-        $logAction->item_type = Asset::class;
-        $logAction->item_id = $asset->id;
-        $logAction->created_at = date('Y-m-d H:i:s');
-        $logAction->created_by = auth()->id();
-        $logAction->logaction('delete');
-    }
 
-    /**
-     * Listen to the Asset deleting event.
-     *
-     * @param  Asset  $asset
-     * @return void
-     */
-    public function restoring(Asset $asset)
-    {
-        $logAction = new Actionlog();
-        $logAction->item_type = Asset::class;
-        $logAction->item_id = $asset->id;
-        $logAction->created_at = date('Y-m-d H:i:s');
-        $logAction->created_by = auth()->id();
-        $logAction->logaction('restore');
-    }
 
     /**
      * Executes every time an asset is saved.
